requirements.txt
# App
frictionless==4.40.3
frictionless[json]
httpx==0.23.0
pytest==7.1.2
python-dotenv==0.20.0

# Tooling
black==22.3.0
isort==5.10.1
flake8==4.0.1
mypy==0.961
pytest-cov==3.0.0
